+++ CHANGELOG.md
@@ -14,9 +14,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   instruments to `NoOpCounter`, `NoOpHistogram`, `NoOpObservableCounter`,
   `NoOpObservableGauge`, `NoOpObservableUpDownCounter`, `NoOpUpDownCounter`
   ([#2616](https://github.com/open-telemetry/opentelemetry-python/pull/2616))
+- Deprecate InstrumentationLibraryInfo and Add InstrumentationScope
+  ([#2583](https://github.com/open-telemetry/opentelemetry-python/pull/2583))
 
 ## [1.11.0-0.30b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.11.0-0.30b0) - 2022-04-18
 
+- Rename API Measurement for async instruments to Observation
+  ([#2617](https://github.com/open-telemetry/opentelemetry-python/pull/2617))
 - Add support for zero or more callbacks
   ([#2602](https://github.com/open-telemetry/opentelemetry-python/pull/2602))
 - Fix parsing of trace flags when extracting traceparent